move rest config loading function to common location
